+++ mgmt/stats/StatType.cc
@@ -783,8 +783,8 @@ StatObject::setTokenValue(StatExprToken * token, bool cluster)
       // only support time function
       token->m_token_value = (StatFloat) (m_current_time - m_last_update);
       if (StatDebug) {
-        Debug(MODULE, "m_current_time(%lld) - m_last_update(%lld) = %lld\n",
-              (long long)m_current_time, (long long)m_last_update, (long long)token->m_token_value);
+        Debug(MODULE, "m_current_time(%"  PRId64 ") - m_last_update(%"  PRId64 ") = %"  PRId64 "\n",
+              (int64_t)m_current_time, (int64_t)m_last_update, (int64_t)token->m_token_value);
       }
       break;
