+++ proxy/logging/Log.cc
@@ -1181,7 +1181,7 @@ Log::flush_thread_main(void *args)
     now = time(NULL);
     if (now > last_time) {
       if ((now % PERIODIC_TASKS_INTERVAL) == 0) {
-        Debug("log-flush", "periodic tasks for %lld", (long long)now);
+        Debug("log-flush", "periodic tasks for %" PRId64, (int64_t)now);
         periodic_tasks(now);
       }
       last_time = (now = time(NULL));
